GEC-468: Port fix for ordering MS files by time during concatenation
MR!426

+++ rapthor/modifystate.py
@@ -134,9 +134,12 @@ def run(parset_file):
                 "images",
                 "visibilities",
             ):
+                output_dir = os.path.join(parset["dir_working"], dirname)
+                if not os.path.exists(output_dir):
+                    continue
                 dcmp = filecmp.dircmp(
                     os.path.join(parset["dir_working"], "pipelines"),
-                    os.path.join(parset["dir_working"], dirname),
+                    output_dir,
                 )
                 for path in (os.path.join(dcmp.right, item) for item in dcmp.right_only):
                     shutil.rmtree(path, ignore_errors=True)
